requirements.txt
flask
python-dotenv
flasgger
langchain_openai
verboselogs
deepgram-sdk
pandas
numpy